+++ controllers/zuul.go
@@ -1016,6 +1016,9 @@ func (r *SFController) EnsureZuulConfigSecret(remoteExecutor bool) *ini.File {
 		ansibleSetupTimeout := int(r.cr.Spec.Zuul.Executor.AnsibleSetupTimeout)
 		cfgINI.Section("executor").NewKey("ansible_setup_timeout", strconv.Itoa(ansibleSetupTimeout))
 	}
+	if r.cr.Spec.Zuul.Executor.Zone != "" {
+		cfgINI.Section("executor").NewKey("zone", r.cr.Spec.Zuul.Executor.Zone)
+	}
 
 	// Enable prometheus metrics
 	for _, srv := range []string{"web", "executor", "scheduler", "merger"} {