+++ docs/reference/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ## [in development]
 
 ### Added
+
+- Zuul.Executor.Zone setting to configure the nodepool executor-zone.
+
 ### Changed
 ### Deprecated
 ### Removed